fix(types): 类型同步优化 fix #12551
