+++ packages/taro-components/types/CoverImage.d.ts
@@ -22,7 +22,6 @@ interface CoverImageProps extends StandardProps {
   referrerPolicy?: 'origin' | 'no-referrer'
 
   /** 点击事件回调。
-   * 版本要求： 基础库 1.9.0 及以上
    * @supported alipay
    */
   onTap?: CommonEventFunction
