+++ packages/taro-components/types/PickerView.d.ts
@@ -58,13 +58,11 @@ interface PickerViewProps extends StandardProps {
   ariaLabel?: string
 
   /** 当滚动选择开始时候触发事件。
-   * 版本要求： 基础库 2.7.23 及以上
    * @supported alipay
    */
   onPickerStart?: CommonEventFunction
 
   /** 当滚动选择结束时候触发事件。
-   * 版本要求： 基础库 2.7.23 及以上
    * @supported alipay
    */
   onPickerEnd?: CommonEventFunction
