+++ packages/taro-components/types/Swiper.d.ts
@@ -119,26 +119,22 @@ interface SwiperProps extends StandardProps {
   disableTouch?: boolean
 
   /** swiper-item 可见时的 class。
-   * 版本要求： 基础库 1.13.7 以及上。
    * @supported alipay
    */
   activeClass?: string
 
   /** acceleration 设置为 {{true}} 时且处于滑动过程中，中间若干屏处于可见时的 class。
-   * 版本要求： 基础库 1.13.7 以及上。
    * @supported alipay
    */
   changingClass?: string
 
   /** 当开启时，会根据滑动速度，连续滑动多屏。
-   * 版本要求： 基础库 1.13.7 及以上
    * @default false
    * @supported alipay
    */
   acceleration?: string
 
   /** 是否禁用代码变动触发 swiper 切换时使用动画。
-   * 版本要求： 基础库 1.13.7 及以上
    * @default false
    * @supported alipay
    */
@@ -150,31 +146,26 @@ interface SwiperProps extends StandardProps {
   onAnimationEnd?: CommonEventFunction<SwiperProps.onCommonEventDetail>
 
   /** 滑动距离阈值，当滑动距离超过阈值时进行 swiper-item 切换。
-   * 版本要求：基础库 1.24.11 及以上
    * @supported alipay
    */
   swipeRatio?: string
 
   /** 滑动综合速度阈值，当超过阈值时进行 swiper-item 切换，数值越小越敏感。
-   * 版本要求：基础库 1.24.11 及以上
    * @supported alipay
    */
   swipeSpeed?: string
 
   /** 计算用户手势时所依赖的滑动角度。角度根据 touchstart 事件和首次 touchmove 事件的坐标计算得出。数值越小越对用户的滑动方向准确度要求越高。
-   * 版本要求：基础库 1.24.11 及以上
    * @supported alipay
    */
   touchAngle?: string
 
   /** 自动以指定滑块的高度为整个容器的高度。当 vertical 为 true 时，默认不调整。可选值为：
-   * 版本要求：基础库 2.6.4 及以上
    * @supported alipay
    */
   adjustHeight?: 'first' | 'current' | 'highest' | 'none'
 
   /** vertical 为 true 时强制使 adjust-height 生效。
-   * 版本要求：基础库 2.6.4 及以上
    * @supported alipay
    */
   adjustVerticalHeight?: string
