+++ packages/taro-components/types/View.d.ts
@@ -34,7 +34,6 @@ interface ViewProps extends StandardProps {
   hoverStayTime?: number
 
   /** 是否阻止区域内滚动页面。
-   * 默认值： false
    * 说明： 如果 view 中嵌套 view，外层 view 设置 disable-scroll 为 true 时禁止内部的滚动。
    * @supported alipay
    * @default false
@@ -42,25 +41,14 @@ interface ViewProps extends StandardProps {
   disableScroll?: string
 
   /** 是否隐藏。
-   * 默认值： false
    * @supported alipay
    * @default false
    */
   hidden?: boolean
 
-  /** 自定义样式名。
-   * @supported alipay
-   */
-  class?: string
-
-  /** 内联样式。
-   * @supported alipay
-   */
-  style?: string
-
   /** 用于动画，详见 my.createAnimation 。使用 my.createAnimation 生成的动画是通过过渡（Transition）实现的，只会触发 onTransitionEnd，不会触发 onAnimationStart, onAnimationIteration, onAnimationEnd。
-   * 默认值：{}
    * @supported alipay
+   * @default {}
    */
   animation?: TaroGeneral.IAnyObject
 
@@ -95,43 +83,36 @@ interface ViewProps extends StandardProps {
   onLongTap?: CommonEventFunction
 
   /** 过渡（Transition）结束时触发。
-   * 版本要求： 基础库 1.8.0 及以上
    * @supported alipay
    */
   onTransitionEnd?: CommonEventFunction
 
   /** 每开启一次新的动画过程时触发。（第一次不触发）
-   * 版本要求： 基础库 1.8.0 及以上
    * @supported alipay
    */
   onAnimationIteration?: CommonEventFunction
 
   /** 动画开始时触发。
-   * 版本要求： 基础库 1.8.0 及以上
    * @supported alipay
    */
   onAnimationStart?: CommonEventFunction
 
   /** 动画结束时触发。
-   * 版本要求： 基础库 1.8.0 及以上
    * @supported alipay
    */
   onAnimationEnd?: CommonEventFunction
 
   /** 当前元素可见面积超过50%时触发。
-   * 版本要求： 基础库 1.9.0 及以上
    * @supported alipay
    */
   onAppear?: CommonEventFunction
 
   /** 当前元素不可见面积超过50%时触发。
-   * 版本要求： 基础库 1.9.0 及以上
    * @supported alipay
    */
   onDisappear?: CommonEventFunction
 
   /** 当前元素首次可见面积达到50%时触发。
-   * 版本要求： 基础库 1.9.4 及以上
    * @supported alipay
    */
   onFirstAppear?: CommonEventFunction
