+++ packages/taro-webpack-runner/src/config/postcss.conf.ts
@@ -1,9 +1,8 @@
 import { isNpmPkg, recursiveMerge } from '@tarojs/helper'
-import { Func } from '@tarojs/runtime'
 import * as path from 'path'
 import { sync as resolveSync } from 'resolve'
 
-import type { IPostcssOption, TogglableOptions } from '@tarojs/taro/types/compile'
+import type { Func, IPostcssOption, TogglableOptions } from '@tarojs/taro/types/compile'
 
 const platform = 'h5'
 const defaultAutoprefixerOption = {
