+++ packages/taro-webpack5-runner/src/plugins/TaroNormalModulesPlugin.ts
@@ -1,9 +1,9 @@
-import { Compiler } from 'webpack'
-
 import TaroSingleEntryDependency from '../dependencies/TaroSingleEntryDependency'
 import { componentConfig } from '../template/component'
-import { Func } from '../utils/types'
 import TaroNormalModule from './TaroNormalModule'
+
+import type { Func } from '@tarojs/taro/types/compile'
+import type { Compiler } from 'webpack'
 
 const walk = require('acorn-walk')
