+++ packages/taro/types/compile/config/index.d.ts
@@ -1,0 +1,6 @@
+export * from './h5'
+export * from './mini'
+
+export * from './manifest'
+export * from './project'
+export * from './util'
