+++ packages/taro/types/compile/index.d.ts
@@ -1,0 +1,3 @@
+export { Current } from '@tarojs/runtime'
+
+export * from './config'
